Adding AI and changing up some variables and functions in the AI section.
Added a custom written Neural network and Matrix libary, and used this
to add a generation based evolution neural network.

The neural network is very simple and does not use crossover to improve
the network. It only uses mutations, which makes it slower.

Everything should work, so this is sort of a release.